Refactor Atelier: move project studio router into studio directory
