--- apps/web/js/views/fire-protection/fireProtection-general.js
@@ -1,26 +1,0 @@
-import { registerProjectPrimaryScrollSource } from "../project-shell-chrome.js";
-
-export function renderFireProtectionGeneral(root) {
-  if (!root) return;
-
-  root.innerHTML = `
-    <section class="settings-section is-active">
-      <div class="settings-card settings-card--param">
-        <div class="settings-card__head">
-          <div>
-            <span class="settings-card__head-title">
-              <h4>Général</h4>
-            </span>
-            <p>Cette section est prête à accueillir les paramètres généraux du domaine Incendie.</p>
-          </div>
-        </div>
-        <div class="settings-placeholder-card">
-          <div class="settings-placeholder-card__title">Placeholder</div>
-          <p>Contenu à définir.</p>
-        </div>
-      </div>
-    </section>
-  `;
-
-  registerProjectPrimaryScrollSource(root.closest("#projectFireProtectionRouterScroll") || document.getElementById("projectFireProtectionRouterScroll"));
-}
